refactor: Reorganize CallListModel
